Move getMixins to BaseClient
almond-cloud's ThingpediaClient is not an HttpClient (it goes
straight to the DB, because it is the backend for the HTTP API)
but it needs the mixins too
