+++ lib/base_client.js
@@ -9,6 +9,8 @@
 // See COPYING for details
 "use strict";
 
+const Mixins = require('./mixins.json');
+
 module.exports = class ThingpediaClientBase {
     constructor(platform) {
         this.platform = platform;
@@ -85,5 +87,12 @@ module.exports = class ThingpediaClientBase {
     /* istanbul ignore next */
     lookupEntity(entityType, searchTerm) {
         throw new Error('not implemented');
+    }
+
+    getMixins() {
+        let mixins = {};
+        for (let mixin of Mixins.data)
+            mixins[mixin.kind] = mixin;
+        return Promise.resolve(mixins);
     }
 };
