+++ lib/http_client.js
@@ -13,7 +13,6 @@ const Tp = require('thingpedia');
 const qs = require('querystring');
 
 const ClientBase = require('./base_client');
-const Mixins = require('./mixins.json');
 
 module.exports = class ThingpediaClientHttp extends ClientBase {
     constructor(platform, url) {
@@ -105,12 +104,5 @@ module.exports = class ThingpediaClientHttp extends ClientBase {
                 array.meta = result.meta;
                 return array;
             });
-    }
-
-    getMixins() {
-        let mixins = {};
-        for (let mixin of Mixins.data)
-            mixins[mixin.kind] = mixin;
-        return Promise.resolve(mixins);
     }
 };
